Remove version from index.js
We'll rely on bower/npm to maintain and check which version of
the library is in use. I would like to include the version on
the nio object on the window but can't find a good way to make
it work with the release tool. I don't want to have to remember
to change that file every time a release is made, but I also can't
think of a good way to have it dynamically reference the package.json
or bower.json files. I'm all ears for suggestions